Add the ability to log check text
Currently when a check is executed, unless the client that is invoking
the check logs the results, it's not possible to know what the result
was. This mitigates that issue by allowing `OkComputer.logger` to be
set to a logger that will log the text when a check is performed. Note
that by default the logger is set to `Logger.new nil`, which performs
a no-op when called, so the current behavior remains unchanged.